requirements.txt
aiohttp==3.7.4
psycopg2-binary==2.8.6
SQLAlchemy==1.4.16
marshmallow==3.12.1
alembic==1.6.5
PyYAML==5.4.1
pytest==6.2.4
asyncpg==0.23.0
pytest_aiohttp==0.3.0
aiohttp-apispec==2.2.1
Faker==8.8.2